collector/pyproject.toml
[project]
name = "mal-collector"
version = "0.1.0"
description = "KRA OpenAPI data collector for mal.kr"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.27",
    "xmltodict>=0.13",
    "sqlalchemy>=2.0",
    "psycopg[binary]>=3.2",
    "python-dotenv>=1.0",
    "typer>=0.12",
    "structlog>=24.1",
    "tenacity>=8.2",
    "feedparser>=6.0",
]

[dependency-groups]
dev = [
    "ruff>=0.6",
    "pytest>=8.0",
]

[tool.uv]
package = false

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP", "B", "SIM"]
